suggested new nav bar
@judsonlmoore

since there isn't a *real* place for a seller to log into yet, let's hide that. you can use a direct link for demos.

and then since it's kind of confusing as a ticket buyer, i got a bit more wordy to call out that they should just click here and not worry about any of the marketing stuff.

i just commented out the old stuff, so feel free to reset if you disagree.